+++ src/constants/index.ts
@@ -90,4 +90,4 @@ export {
 } from "@src/constants/tour.constants";
 export { getArrowStyles } from "@src/constants/dashedArrowStyles.constants";
 export { ActivityState } from "@src/constants/activities.constants";
-export { getBillingPlanFeatures, type BillingPlanFeature } from "@constants/lists";
+export { getBillingPlanFeatures } from "@constants/lists";
