+++ src/constants/lists/index.ts
@@ -7,4 +7,4 @@ export {
 	selectIntegrationGoogle,
 	selectIntegrationAws,
 } from "@constants/lists/connections";
-export { getBillingPlanFeatures, type BillingPlanFeature } from "@constants/lists/billingPlanFeatures.constants";
+export { getBillingPlanFeatures } from "@constants/lists/billingPlanFeatures.constants";
