+++ src/interfaces/components/apexMinMaxChart.interface.ts
@@ -1,0 +1,4 @@
+export interface ApexMinMaxGradientCircleChart {
+	max: number;
+	value: number;
+}
