+++ src/interfaces/components/billing.interface.ts
@@ -1,8 +1,3 @@
-export interface HalfCircleProgressBarProps {
-	max: number;
-	value: number;
-}
-
 export interface BillingFeature {
 	name: string;
 	free: string;
