+++ src/interfaces/models/billing.interface.ts
@@ -27,16 +27,6 @@ export interface Usage {
 	usage: UsageItem[];
 }
 
-export interface CheckoutSessionRequest {
-	stripe_price_id: string;
-	success_url: string;
-}
-
-export interface CheckoutSessionResponse {
-	redirectUrl: string;
-	sessionId: string;
-}
-
 export interface BillingPlanFeature {
 	name: string;
 	free: string;
