+++ src/interfaces/models/index.ts
@@ -13,3 +13,11 @@ export type {
 	SessionOutputLog,
 } from "@src/interfaces/models/session.interface";
 export type { UserStatusKeyType } from "@src/interfaces/models/userStatus.interface";
+export type {
+	BillingPlanFeature,
+	PaymentOption,
+	Plan,
+	PlanLimit,
+	Usage,
+	UsageItem,
+} from "@src/interfaces/models/billing.interface";
