+++ src/interfaces/services/index.ts
@@ -1,0 +1,4 @@
+export type {
+	CheckoutSessionRequest,
+	CheckoutSessionResponse,
+} from "@src/interfaces/services/billingService.interface";
