+++ src/interfaces/store/billingStore.interface.ts
@@ -1,0 +1,12 @@
+import { Plan, Usage } from "../models";
+
+export interface BillingStoreState {
+	plans: Plan[];
+	usage?: Usage;
+}
+
+export interface BillingLoadingState {
+	billing: boolean;
+	plans: boolean;
+	usage: boolean;
+}
