+++ src/types/index.ts
@@ -7,4 +7,3 @@ export type { StoreCallbacks } from "@type/factories.type";
 export type { TourStepKeys, TourStepValues, IndexedDBTourType } from "@type/tour.type";
 export type { ActivityStateType } from "@type/models";
 export * from "./iframeCommunication.type";
-export * from "./billing.types";
