+++ src/types/stores/organizationsStore.types.ts
@@ -1,6 +1,6 @@
 import { MemberStatusType, UserStatusType } from "@src/enums";
+import { Plan, Usage } from "@src/interfaces/models";
 import { ServiceResponse } from "@src/types";
-import { Plan, Usage } from "@src/types/billing.types";
 import { EnrichedOrganization, Organization, OrganizationMember, EnrichedMember, User } from "@type/models";
 import { StoreResponse } from "@type/stores";
