+++ tsconfig.json
@@ -50,7 +50,6 @@
 			"@i18n/*": ["./src/i18n/*"],
 			"@i18n": ["./src/i18n/index"],
 			"@interfaces/*": ["./src/interfaces/*"],
-			"@interfaces": ["./src/interfaces/index"],
 			"@type/*": ["./src/types/*"],
 			"@type": ["./src/types/index"],
 			"@pages/*": ["./pages/*"],
